ci: build docker image on PRs / tags (#2353)
<!--
Please read and fill out this form before submitting your PR.

Please make sure you have reviewed our contributors guide before
submitting your
first PR.

NOTE: PR titles should follow semantic commits:
https://www.conventionalcommits.org/en/v1.0.0/
-->

## Overview

This PR adds a workflow which will build the rollkit image on each PR,
in a follow up we can use this image to perform upgrade tests.

If it's on a PR, it will be tagged like `pr-1234`, if it's a tag or main
it will be tagged with the format `v1.0.0` or `main`.

<!-- 
Please provide an explanation of the PR, including the appropriate
context,
background, goal, and rationale. If there is an issue with this
information,
please provide a tl;dr and link the issue. 

Ex: Closes #<issue number>
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Updated continuous integration workflows to support dynamic Docker
image tagging for test jobs.
- Added automated building and publishing of Docker images during
testing.
  - Introduced a placeholder for future upgrade tests.
- Removed the separate Docker build and publish workflow to streamline
processes.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->